refactor: separate config.options out from init.lua
